ci: split into smaller workflows with some improvements (#1475)
## Summary
This PR splits our big CI workflow into smaller reusable workflows,
enabling a greater degree of parallelization and improve
maintainability.

## Details
Previously, the main workflow runs in stages for each of the tested
platforms. However, this poses a problem where a stage is blocked by the
slowest runner and none of the other runners could move on.

With the release of reusable workflows for Actions, we can now separate
the stages into a different workflow, then parallelize that instead,
allowing faster runners to proceed through test stages ahead of the
rest. The response time for test results should now be greatly improved,
with the M1 runner finishing its run in just 7 minutes.

This separation also opened the doors to customizing how tests are run
for each platforms, and we now only use 1 test batch for *nix while
allocating 3 to Windows. This normalizes test time between Windows and
*nix, while allowing freed up runners to process other jobs.

Lastly, some tiny QoL improvements are added to the pipeline:

- Test failures will now be displayed at the end of "Run tester" step,
reducing the number of clicks required to see all failures to 1.
- The diffoscope container is now used instead of the Ubuntu package,
reducing installation time and let us receive latest features/bug
fixes.
- If there is a diffoscope failure, a test summary is now generated with
a link to the test report. This should make it a lot easier to get to
the report should that be necessary.