fix memory corruption when using standalone heap (#1478)
## Summary

Fix using the standalone heap (enabled with `--os:standalone` or
`-d:StandaloneHeapSize`) resulting in memory corruption due to
incorrect page alignment.

## Details

The initial pointer for the array-based page allocator is now aligned
to the size of a page, which ensures that `osAllocPages` always returns
a page-aligned pointer (the default allocator relies on this to be the
case).

A test is added to ensure that simple allocation/deallocation works
when using the array-based page allocator.