fix(sem): crash with erroneous `is` expression (#1485)
## Summary

Fix a crash with erroneous `is` expressions appearing directly as a
`when` conditions expressions causing the compiler to crash.

## Details

There were two issues:
1. `semIs` modified input AST, which led to the error correction in the
   `fitNode` called by `forceBool` called by `semWhen` not triggering
2. `semWhen` used the `semConstExpr`, which doesn't return an `nkError`
   on failure, thus leading to a field access defect when happening in
   a context where `localReport` doesn't terminate the compiler

`semIs` is changed to not modify input AST. In addition, the workaround
for `is` being analyzed eagerly in generic expression is removed, as
it's no longer needed.

In `semWhen`, the usage of `semConstExpr` is replaced with
`semRealConstExpr`, which properly propagates `nkError`. The latter was
missing execution context handling (needed for detecting illegal
`break`s crossing the compile-time / run-time boundary), which is now
fixed too.

Fixes https://github.com/nim-works/nimskull/issues/1483.